Update: Modular displays with corresponding options windows (WIP)
This update modularizes the current display so that you can
dynamically add and remove different kinds of displays. Each display
comes with a corresponding state. Each display appears for every core

Displays can be repeating (multiple optical/element-count displays)

This currently disables the element selection list (which was ugly
anyway..). This will be replaced entirely in a future update