Update: Refactor to use Contexts/Subscriptions
This is almost a complete rewrite..

The benefit is two fold:

- All Lifecycles/Effects in effect have access to the whole state from
everywhere without having to actually pass tons of data around. As
dependencies dynamically evolve with product requirements then
subscriptions are simply added. If some component suddenly needs
access to a state variable you no longer need to refactor the heirarchy
to make sure it get that variable passed in.

- Subscriptions use memoization - so state variables that are
derivatives of other variables are now stores in a hidden cache that
automatically is updated as the underlying variables change. So you
don't need to worry about updating your state. Changes cascade and
propagate naturally. This *greatly* simplifies logic

Thanks for the help and guiadance from vlaaad:
https://github.com/cljfx/cljfx/issues/69