Added: saving xrf-scan to file with click (ugly)
Ugly:
- I no longer use a set b/c I need to preserve the column order. It
just looks a bit weird if the output has scrambled columns. But the
code excepts a set.. so it's a bit ugly. Maybe I need a wrapper or
something
- The click to save is in an outer scope (it wouldn't trigger when
attached to the image-view) - so in theory it can trigger when you
have no xrf-scan loaded in. However in my tests it doesn't trigger
when the button is there. So maybe it's okay... for now.. I don't have
a good alternative for now