Added: can select plottable elements that aren't in the first core
The plottable? check was only checking if the element was in the first
core and if it was a Long. Now if it doesn't find the element at
all (ie. it's returning nil) then it'll go to the next core and check
if it's there

If you do something crazy like have one XRF file have numbers in a
columns and then another column has strings for the same column
name.. well then you're screwed. But then the user is crazy