Update: Correlate only over half length
Correlating over the full length makes things go wonky b/c in the
former half the cores will switch in depth relative to each other. So
then you end up adjusting the depth of a core than is now "higher up"
and this isn't how the software is set up. The deeper cores is the one
that is always the one being adjusted